Extract out utils and tidy index.js
