+++ index.js
@@ -6,9 +6,11 @@ import uid from 'https://cdn.pika.dev/uid'
 
 const html = htm.bind(h)
 
-import Tests from '../../components/tests.js'
-import Suites from '../../components/suites.js'
-import Results from '../../components/results.js'
+import Tests from './components/tests.js'
+import Archive from './components/archive.js'
+import Results from './components/results.js'
+
+import { pSeries, average, toURL } from './utils.js'
 
 const defaults = {
   started: false,
@@ -19,8 +21,8 @@ const defaults = {
   duration: 3,
   progress: 0,
   id: uid(),
-  title: 'Finding numbers in an array',
-  before: `const data = [...Array(800).keys()]`,
+  title: 'Finding numbers in an array of 1000',
+  before: `const data = [...Array(1000).keys()]`,
   tests: [
     { name: 'Find item 2000', code: 'data.find(x => x == 200)', ops: 0 },
     { name: 'Find item 4000', code: 'data.find(x => x == 400)', ops: 0 },
@@ -41,67 +43,19 @@ const reducer = (state, update) => ({
   ...(typeof update === 'function' ? update(state) : update),
 })
 
-const pReduce = (iterable, reducer, initialValue) =>
-  new Promise((resolve, reject) => {
-    const iterator = iterable[Symbol.iterator]()
-    let index = 0
+const startTesting = state => ({
+  tests: state.tests.map(test => ({ ...test, ops: 0 })),
+  started: true,
+  progress: 0,
+})
 
-    const next = async total => {
-      const element = iterator.next()
+const latestLocalStorage = () => ({
+  suites: Object.entries(localStorage).map(([k, v]) => [k, JSON.parse(v)]),
+})
 
-      if (element.done) {
-        resolve(total)
-        return
-      }
-
-      try {
-        const value = await Promise.all([total, element.value])
-        next(reducer(value[0], value[1], index++))
-      } catch (error) {
-        reject(error)
-      }
-    }
-
-    next(initialValue)
-  })
-
-const pSeries = async tasks => {
-  const results = []
-
-  await pReduce(tasks, async (_, task) => {
-    const value = await task()
-    results.push(value)
-  })
-
-  return results
-}
-
-function average(arr) {
-  var sums = {},
-    counts = {},
-    results = [],
-    ids = {},
-    name
-  for (var i = 0; i < arr.length; i++) {
-    name = arr[i].code
-    if (!(name in sums)) {
-      sums[name] = 0
-      counts[name] = 0
-      ids[name] = arr[i].name
-    }
-    sums[name] += arr[i].ops
-    counts[name]++
-  }
-
-  for (name in sums) {
-    results.push({
-      name: ids[name],
-      code: name,
-      ops: (sums[name] / counts[name]) << 0,
-    })
-  }
-  return results
-}
+const updateProgress = state => ({
+  progress: state.progress + state.tests.length,
+})
 
 const app = ({ WORKER }) => {
   const [state, dispatch] = useReducer(reducer, init)
@@ -119,23 +73,22 @@ const app = ({ WORKER }) => {
 
   useEffect(() => {
     if (started) {
+      const bench = test => () =>
+        new Promise(resolve => {
+          const worker = new Worker(WORKER)
+          worker.onmessage = e => {
+            const ops = (e.data * (1000 / duration)) << 0
+            resolve({ ...test, ops })
+            worker.terminate()
+          }
+          worker.postMessage([before, test, duration])
+        })
+
       const tasks = () => () => {
-        const run = pSeries(
-          tests.map(test => () =>
-            new Promise(resolve => {
-              const worker = new Worker(WORKER)
-              worker.onmessage = e => {
-                const ops = (e.data * (1000 / duration)) << 0
-                resolve({ ...test, ops })
-                worker.terminate()
-              }
-              worker.postMessage([before, test, duration])
-            })
-          )
-        )
-        dispatch(state => ({ progress: state.progress + tests.length }))
-        return run
+        dispatch(updateProgress)
+        return pSeries(tests.map(bench))
       }
+
       pSeries(Array.from({ length: runs }, tasks)).then(results => {
         dispatch({ tests: average(results.flat()), started: false })
       })
@@ -143,17 +96,11 @@ const app = ({ WORKER }) => {
   }, [started, tests])
 
   useEffect(() => {
-    const x = JSON.stringify({ title, before, tests, updated: new Date(), id })
+    const x = JSON.stringify({ id, title, before, tests, updated: new Date() })
     history.replaceState(null, null, `#${encodeURIComponent(btoa(x))}`)
     if (Object.fromEntries(suites)[id]) {
-      console.log(before, JSON.stringify(tests))
       localStorage.setItem(id, x)
-      dispatch({
-        suites: Object.entries(localStorage).map(([k, v]) => [
-          k,
-          JSON.parse(v),
-        ]),
-      })
+      dispatch(latestLocalStorage)
     }
   }, [id, title, before, tests])
 
@@ -166,11 +113,7 @@ const app = ({ WORKER }) => {
           e.keyCode == 13
         ) {
           e.preventDefault()
-          dispatch(state => ({
-            tests: state.tests.map(test => ({ ...test, ops: 0 })),
-            started: true,
-            progress: 0,
-          }))
+          dispatch(startTesting)
         }
       },
       false
@@ -185,7 +128,7 @@ const app = ({ WORKER }) => {
             <${Results} state=${state} dispatch=${dispatch} />
           `
         : html`
-            <${Suites} state=${state} dispatch=${dispatch} />
+            <${Archive} state=${state} dispatch=${dispatch} />
           `}
       ${dialog &&
         html`
@@ -196,21 +139,13 @@ const app = ({ WORKER }) => {
               <br />
               Reliably compare code exectution times in browser.
             </p>
-            <button
-              onClick=${_ => {
-                dispatch({ dialog: false, started: true })
-              }}
-            >
+            <button onClick=${() => dispatch({ dialog: false, started: true })}>
               Start Experimenting
             </button>
           </dialog>
         `}
     </main>
   `
-}
-
-function toURL(code, type = 'application/javascript') {
-  return URL.createObjectURL(new Blob([code], { type }))
 }
 
 fetch('./run.js')
